+++ app/src/main/java/com/example/wallet/data/util/Constants.kt
@@ -1,0 +1,5 @@
+package com.example.wallet.data.util
+
+object Constants {
+    const val DB_NAME = "Expenses_db"
+}
