+++ app/src/main/java/com/example/wallet/data/util/preferences/application/ApplicationPreferences.kt
@@ -1,4 +1,4 @@
-package com.example.wallet.ui.util.preferences
+package com.example.wallet.data.util.preferences.application
 
 import kotlinx.coroutines.flow.Flow
