+++ app/src/main/java/com/example/wallet/presentation/component/main/MainScaffold.kt
@@ -1,4 +1,4 @@
-package com.example.wallet.ui.component.main
+package com.example.wallet.presentation.component.main
 
 import android.content.res.Configuration
 import androidx.compose.foundation.layout.PaddingValues
@@ -14,8 +14,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.wallet.R
-import com.example.wallet.ui.component.base.AppBaseSurface
-import com.example.wallet.ui.model.FabActionButton
+import com.example.wallet.presentation.component.base.AppBaseSurface
+import com.example.wallet.presentation.model.FabActionButton
 
 @Preview(showBackground = true, name = "Light Mode")
 @Preview(showBackground = true, name = "Dark Mode", uiMode = Configuration.UI_MODE_NIGHT_YES)
