+++ app/src/main/java/com/example/wallet/presentation/component/reusable/Transformations.kt
@@ -1,4 +1,4 @@
-package com.example.wallet.ui.component.reusable
+package com.example.wallet.presentation.component.reusable
 
 import android.util.Log
 import androidx.compose.ui.text.AnnotatedString
