+++ app/src/main/java/com/example/wallet/presentation/screen/HomeScreen.kt
@@ -1,4 +1,4 @@
-package com.example.wallet.ui.screen
+package com.example.wallet.presentation.screen
 
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
 import androidx.compose.foundation.clickable
@@ -21,8 +21,8 @@ import androidx.navigation.compose.rememberNavController
 import com.example.wallet.R
 import com.example.wallet.data.entity.Expense
 import com.example.wallet.data.viewmodel.HomeViewModel
-import com.example.wallet.ui.component.base.AppBaseSurface
-import com.example.wallet.ui.navigation.Screen
+import com.example.wallet.presentation.component.base.AppBaseSurface
+import com.example.wallet.navigation.Screen
 
 @Composable
 @Preview(showBackground = true, name = "Light Mode")
